# src/main/java/com/mindSpring/swagger/swagger.yaml
...
openapi: 3.0.1
info:
  title: Swagger
  description: REST API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  "/member/{memberId}/category/{categoryId}/worry":
    get:
      tags:
        - 고민 관리
      summary: 고민 전체 조회(고민 목록 페이지)
      description: 고민 전체 정보를 조회합니다.
      operationId: getWorries
      parameters:
        - name: memberId
          in: path
          description: 회원 ID
          required: true
          schema:
            type: integer
            format: int64
        - name: categoryId
          in: path
          description: 카테고리 ID
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          description: page
          required: true
          schema:
            type: integer
            format: int32
        - name: size
          in: query
          description: size
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 조회 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/WorryResponseDto"
        '400':
          description: 회원을 찾을 수 없음
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: 카테고리를 찾을 수 없음
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
    post:
      tags:
        - 고민 관리
      summary: 고민 등록(고민 작성 페이지)
      description: 새로운 고민을 등록합니다.
      operationId: createWorry
      parameters:
        - name: memberId
          in: path
          description: 회원 ID
          required: true
          schema:
            type: integer
            format: int64
        - name: categoryId
          in: path
          description: 카테고리 ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/WorryRequestDto"
        required: true
      responses:
        '200':
          description: 등록 성공
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: 회원을 찾을 수 없음
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: 카테고리를 찾을 수 없음
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...
  "/member/{memberId}/category/{categoryId}/worry/{worryId}":
    get:
      tags:
        - 고민 관리
      summary: 고민 조회(고민 상세 페이지)
      description: 고민 ID로 고민 정보를 조회합니다.
      operationId: getWorry
      parameters:
        - name: memberId
          in: path
          description: 회원 ID
          required: true
          schema:
            type: integer
            format: int64
        - name: categoryId
          in: path
          description: 카테고리 ID
          required: true
          schema:
            type: integer
            format: int64
        - name: worryId
          in: path
          description: 고민 ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 조회 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/WorryResponseDto"
        '400':
          description: 회원을 찾을 수 없음
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: 카테고리를 찾을 수 없음
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '402':
          description: 고민를 찾을 수 없음
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    WorryRequestDto:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
      description: 고민 정보
...
    WorryResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time